Wallet class removed and logic moved to the Accounts module. Proxy implemented for the wallet property in the Accounts module to not break any public API
