+++ packages/web3-eth-accounts/tests/src/models/WalletTest.js
@@ -1,0 +1,25 @@
+import Wallet from '../../../src/models/Wallet';
+
+// Mocks
+jest.mock('');
+
+/**
+ * Wallet test
+ */
+describe('WalletTest', () => {
+    let wallet;
+
+    beforeEach(() => {
+        wallet = new Wallet();
+    });
+
+    it('constructor check', () => {
+        expect(wallet.length).toEqual(0);
+
+        expect(wallet.defaultKeyName).toEqual('web3js_wallet');
+    });
+
+    it('calls findSafeIndex and returns the expected pointer', () =< {
+
+    })
+});
